pyproject.toml
[build-system]
requires = ["setuptools"]
build-meta = "setuptools.build_meta"

[project]
name = "DJGaro"
version = "1.0.0"
description = "Minimalistic discord bot which streams music from youtube"
readme = "README.md"
dependencies = [
	"discord.py",
	"validators",
	"isoduration",
	"dotenv",
]

[project.scripts]
garo-start = "djgaro.main:main"

[tool.setuptools.packages.find]
exclude = ["test*", "testing*"]
namespaces = false
